+++ Runtime/Model/BacktraceReport.cs
@@ -223,11 +223,11 @@ namespace Backtrace.Unity.Model
         /// </summary>
         internal void SetReportFingerPrintForEmptyStackTrace()
         {
-            if (string.IsNullOrEmpty(Exception.StackTrace))
+            if (Exception != null && string.IsNullOrEmpty(Exception.StackTrace))
             {
                 // set attributes instead of fingerprint to still allow our user to define customer
                 // fingerprints for reports without stack trace and apply deduplication rules in report flow.
-                Attributes["_mod_fingerprint"] = Exception.Message.OnlyLetters().GetSha();
+                Attributes["_mod_fingerprint"] = Message.OnlyLetters().GetSha();
             }
 
         }